refactor to search component; update tests
