+++ lib/railway_ui_web/templates/published_message/index.html.leex
@@ -2,7 +2,7 @@
 <div class="row justify-content-md-center">
   <div class="col-md-auto">
     <h1 class="display-3">Published Messages</h1>
-    <%= render RailwayUiWeb.MessageView, "search.html", assigns %>
+    <%= live_component @socket, RailwayUiWeb.MessageLive.SearchComponent, view: @view, id: :search %>
     <table class="table table-responsive pt-3 table-hover" phx-update="replace">
       <thead>
         <tr>
