+++ test/railway_ui_web/live/consumed_message_live/index_test.exs
@@ -70,8 +70,8 @@ defmodule RailwayUiWeb.ConsumedMessageLive.IndexTest do
       end)
 
       assert html != Enum.at(consumed_messages, 3).uuid
-      html = render_submit(view, :search, %{"search" => %{"query" => query_filter, "value" => query_value}})
-      assert html =~ Enum.at(consumed_messages, 3).correlation_id
+      send(view.pid, {:search, %{"search" => %{"query" => query_filter, "value" => query_value}}})
+      assert render(view) =~ Enum.at(consumed_messages, 3).correlation_id
     end
 
     test "it searches by correlation ID", %{html: html, view: view, consumed_messages: consumed_messages} do
@@ -90,8 +90,8 @@ defmodule RailwayUiWeb.ConsumedMessageLive.IndexTest do
       end)
 
       assert html != Enum.at(consumed_messages, 3).uuid
-      html = render_submit(view, :search, %{"search" => %{"query" => query_filter, "value" => query_value}})
-      assert html =~ Enum.at(consumed_messages, 3).uuid
+      send(view.pid, {:search, %{"search" => %{"query" => query_filter, "value" => query_value}}})
+      assert render(view) =~ Enum.at(consumed_messages, 3).uuid
     end
 
     test "it searches by message type with pagination", %{html: html, view: view, consumed_messages: consumed_messages} do
@@ -110,7 +110,7 @@ defmodule RailwayUiWeb.ConsumedMessageLive.IndexTest do
       end)
 
       assert html != Enum.at(consumed_messages, 3).uuid
-      html = render_submit(view, :search, %{"page" => "2", "search" => %{"query" => query_filter, "value" => query_value}})
-      assert html =~ Enum.at(consumed_messages, 3).uuid
+      send(view.pid, {:search, %{"page" => "2", "search" => %{"query" => query_filter, "value" => query_value}}})
+      assert render(view) =~ Enum.at(consumed_messages, 3).uuid
     end
 end
